Brathna

+++ app/Http/Controllers/UploadPhotoController.php
@@ -30,6 +30,14 @@ class UploadPhotoController extends Controller
 
     public function store(Request $request)
     {
+        // Validate the request data
+        $request->validate([
+            'title' => 'required|string|max:70',
+            'description' => 'required|string|max:800',
+            'image' => 'required|url',
+            'location' => 'required|string'
+        ]);
+
         $post = new Post();
         $post->title = $request->input('title');
         $post->description = $request->input('description');
@@ -59,6 +67,9 @@ class UploadPhotoController extends Controller
 
     public function edit(Post $post)
     {
+        // Check if the user is authorized to edit the post
+        $this->authorize('update', $post);
+    
         return view('uploadphoto.edit', [
             'post' => $post
         ]);
@@ -66,7 +77,19 @@ class UploadPhotoController extends Controller
 
     public function update(Request $request, Post $post)
     {
-        $post->update($request->only('title', 'description', 'image', 'location'));
+        // Validate the request data
+        $request->validate([
+            'title' => 'required|string|max:70',
+            'description' => 'required|string|max:800',
+            'image' => 'required|url',
+            'location' => 'required|string'
+        ]);
+
+        // Fill the post attributes with the request data
+        $post->fill($request->all());
+
+        // Save the updated post to the database
+        $post->save();
 
         return redirect()->route('profile.index')
             ->withSuccess(__('Post updated successfully.'));
@@ -74,6 +97,9 @@ class UploadPhotoController extends Controller
 
     public function destroy(Post $post)
     {
+        // Check if the user is authorized to edit the post
+        $this->authorize('update', $post);
+        
         $post->delete();
 
         return redirect()->route('profile.index')
